fix: id types from int to long

+++ src/main/java/com/example/curator/model/Artist.java
@@ -17,8 +17,8 @@ import lombok.NoArgsConstructor;
 public class Artist {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
     private String apiOrigin;
-    private int apiID;
+    private Long apiID;
 }
